Fix warning.

+++ source/gui/theme/imgui_theme.cpp
@@ -994,7 +994,7 @@ namespace ImGuiTheme
     bool _ShowThemeSelector(ImGuiTheme_* theme)
     {
         bool changed = false;
-        ImVec2 listboxSize =  ImVec2(0.f, ImGuiTheme_Count * (ImGui::GetFontSize() + ImGui::GetStyle().ItemInnerSpacing.y) );
+        ImVec2 listboxSize =  ImVec2(0.f, static_cast<int>(ImGuiTheme_Count) * (ImGui::GetFontSize() + ImGui::GetStyle().ItemInnerSpacing.y) );
         if (ImGui::BeginListBox("Available_themes", listboxSize))
         {
             for (int i = 0; i < ImGuiTheme_Count; ++i)
